+++ tests/cli/clitest.cpp
@@ -1434,6 +1434,17 @@ Options:
 
   --help     Show this message and exit.
 )");
+
+        notOn.defaultDesc({});
+        EXPECT_HELP(cli, "", 1 + R"(
+Usage: test [OPTIONS]
+
+Options:
+  / --notOn
+  --on
+
+  --help     Show this message and exit.
+)");
     }
 
     {